+++ packages/examples/main/src/utils/fs-utils.ts
@@ -2,9 +2,13 @@
  * Copyright (c) 2018-2022 TypeFox GmbH (http://www.typefox.io). All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-import { StreamMessageReader, StreamMessageWriter } from 'vscode-jsonrpc/node.js';
-import { start } from './json-server.js';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
 
-const reader = new StreamMessageReader(process.stdin);
-const writer = new StreamMessageWriter(process.stdout);
-start(reader, writer);
+/**
+ * Solves: __dirname is not defined in ES module scope
+ */
+export const getLocalDirectory = (referenceUrl: string | URL) => {
+    const __filename = fileURLToPath(referenceUrl);
+    return dirname(__filename);
+};